% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aSPC_dcor2.R
\name{aSPC_dcor2}
\alias{aSPC_dcor2}
\title{An Adaptive Sum of Powered Correlation Test (aSPC) with dcor}
\usage{
aSPC_dcor2(df1, df2, pow = pow, B = B)
}
\arguments{
\item{df1, }{first matrix}

\item{df2, }{second matrix}

\item{pow, }{power integer candidates, default c(1:8, Inf)}

\item{B, }{number of permutations to calculate a P-value}
}
\value{
the P-values of SPC and aSPC tests
}
\description{
An Adaptive Sum of Powered Correlation Test (aSPC) with dcor
}
\references{
Xu Z., Pan W. An adaptive and powerful test for two groups of variables with high dimension
}
\keyword{internal}

