% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcor_list2.R
\name{dcor_list2}
\alias{dcor_list2}
\title{calculate the correlation matrix}
\usage{
dcor_list2(ls_distCenterMat_X, ls_distCenterMat_Y, ls_distvariance_X,
  ls_distvariance_Y)
}
\arguments{
\item{ls_distCenterMat_X, }{first list of doubly centered distance matrices for p columns in df1}

\item{ls_distCenterMat_Y, }{second list of doubly centered distance matrices for p columns in df1}

\item{ls_distvariance_X, }{first list of distance variance in df1}

\item{ls_distvariance_Y, }{second list of distance variance in df2}
}
\value{
a p by q correlation matrix
}
\description{
calculate the correlation matrix
}
\keyword{internal}

