% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat.R
\name{dcor_mat}
\alias{dcor_mat}
\title{calculate the correlation matrix by dcor() in package energy}
\usage{
dcor_mat(df1, df2)
}
\arguments{
\item{df1, }{first matrix}

\item{df2, }{second matrix}
}
\value{
a p by q correlation matrix
}
\description{
calculate the correlation matrix by dcor() in package energy
}
\keyword{internal}

