% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doublyCenterDist.R
\name{get_doublyCenterDist}
\alias{get_doublyCenterDist}
\title{my program to get doubly center distance matrix}
\usage{
get_doublyCenterDist(x, index = 1)
}
\arguments{
\item{x, }{distance matrix or data matrix}

\item{index, }{power to the Euclidean norm}
}
\value{
the P-values of SPC and aSPC tests
}
\description{
my program to get doubly center distance matrix
}
\references{
Xu Z., Pan W. An adaptive and powerful test for two groups of variables with high dimension
}
\keyword{internal}

