% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_DCOR.R
\name{My_DCOR}
\alias{My_DCOR}
\title{my program to get distance correlation}
\usage{
My_DCOR(A, B, dVarX, dVarY)
}
\arguments{
\item{A, }{doubly centered matrix from get_doublyCenterDist(x, index)$mat}

\item{B, }{doubly centered matrix from get_doublyCenterDist(x, index)$mat}

\item{dVarX, }{doubly centered matrix from get_doublyCenterDist(x, index)$dVar}

\item{dVarY, }{doubly centered matrix from get_doublyCenterDist(x, index)$dVar}
}
\value{
the P-values of SPC and aSPC tests
}
\description{
my program to get distance correlation
}
\references{
Xu Z., Pan W. An adaptive and powerful test for two groups of variables with high dimension
}
\keyword{internal}

